requirements.txt
# requirements.txt
streamlit
google-generativeai>=0.4.0
requests
python-dotenv